Entry level cradle-to-grave exploit Lab
This Lab was written to be the instructors guidelines when demo-ing.  Assuming system access, this lab starts with a blue team tool, OpenVAS.  It then quickly pivots to how a vulnerability scan can be used by an attacker to understand and exploit vulnerabilities.  At the Security+ level the focus is how to correct the vulnerabilities that OpenVAS finds.  At the Pentest level, we walk through nmap enumeration, exploit-db and NVD correlation, exploring the Metasploit Framework on a Kali machine to use a preloaded script for an exploit AND how to read that script to perform that attack manually.  Once we have a shell, we will extract usernames and passwords and use John the Ripper to crack the hashes. Finally, we show the difference between a Reverse Shell and a Bind shell with a discussion of how the Social Engineering Toolkit can be used to put the pieces in place for a reverse or bind shell.